tweak composer versioning per semver
Make the require section a bit more strict to avoid suprises,
but let the require-dev section update up to next major version.

This assumes that all packages respect semver, but in any case
our composer.json is in the repository too so we should be fine.

See https://getcomposer.org/doc/01-basic-usage.md#next-significant-release-tilde-and-caret-operators-