Update api docs
User model has changed - username instead of email and name. Age is now encrypted
and removed in some places, but did not change it here yet.

Required field can not have a default defined, and if something is a boolean
it expects true/false. Change those to integer for now, let me know.